+++ src/BedLowCoverage/main.cpp
@@ -27,9 +27,9 @@ public:
 		addOutfile("out", "Output BED file. If unset, writes to STDOUT.", true);
 		addInt("min_mapq", "Minimum mapping quality to consider a read.", true, 1);
 		addInt("min_baseq", "Minimum base quality to consider a base.", true, 0);
-		addString("ref", "Reference genome for CRAM compression (compulsory for CRAM support).", true);
+		addString("ref", "Reference genome for CRAM support (mandatory if CRAM is used).", true);
 
-		changeLog(2020,  11, 27, "Added Cram support.");
+		changeLog(2020,  11, 27, "Added CRAM support.");
 		changeLog(2020,  5,  26, "Added parameter 'min_baseq'.");
 		changeLog(2016,  6,  9, "The BED line name of the input BED file is now passed on to the output BED file.");
 	}