+++ src/GSvar/MainWindow.cpp
@@ -27,6 +27,7 @@
 #include <QChartView>
 #include <GenLabDB.h>
 #include <QToolTip>
+#include <QProcess>
 QT_CHARTS_USE_NAMESPACE
 #include "ReportWorker.h"
 #include "ScrollableTextDialog.h"
@@ -111,6 +112,7 @@ QT_CHARTS_USE_NAMESPACE
 #include "DiseaseCourseWidget.h"
 #include "CfDNAPanelWidget.h"
 #include "ClinvarSubmissionGenerator.h"
+#include "AlleleBalanceCalculator.h"
 
 MainWindow::MainWindow(QWidget *parent)
 	: QMainWindow(parent)
@@ -617,6 +619,13 @@ void MainWindow::on_actionShowPublishedVariants_triggered()
 	dlg->exec();
 }
 
+void MainWindow::on_actionAlleleBalance_triggered()
+{
+	AlleleBalanceCalculator* widget = new AlleleBalanceCalculator();
+	auto dlg = GUIHelper::createDialog(widget, "Allele balance of heterzygous variants");
+	dlg->exec();
+}
+
 void MainWindow::on_actionClose_triggered()
 {
 	loadFile();
@@ -630,14 +639,10 @@ void MainWindow::on_actionCloseMetaDataTabs_triggered()
 	}
 }
 
-void MainWindow::on_actionIgvInit_triggered()
-{
-	igv_initialized_ = false;
-}
-
 void MainWindow::on_actionIgvClear_triggered()
 {
 	executeIGVCommands(QStringList() << "new");
+	igv_initialized_ = false;
 }
 
 void MainWindow::on_actionIgvPort_triggered()
@@ -1182,8 +1187,8 @@ void MainWindow::handleInputFileChange()
 
 void MainWindow::variantCellDoubleClicked(int row, int /*col*/)
 {
-	int var_index = ui_.vars->rowToVariantIndex(row);
-	openInIGV(variants_[var_index].toString());
+	const Variant& v = variants_[ui_.vars->rowToVariantIndex(row)];
+	openInIGV(v.chr().str() + ":" + QString::number(v.start()) + "-" + QString::number(v.end()));
 }
 
 void MainWindow::variantHeaderDoubleClicked(int row)
@@ -1194,98 +1199,96 @@ void MainWindow::variantHeaderDoubleClicked(int row)
 	editVariantReportConfiguration(var_index);
 }
 
-void MainWindow::openInIGV(QString region)
+bool MainWindow::initializeIvg(QAbstractSocket& socket)
 {
-	QStringList init_commands;
-	if (!igv_initialized_)
+	IgvDialog dlg(this);
+
+	//sample VCF
+	QString folder = QFileInfo(filename_).absolutePath();
+	QStringList files = Helper::findFiles(folder, "*_var_annotated.vcf.gz", false);
+
+	if (files.count()==1)
 	{
-		IgvDialog dlg(this);
+		QString name = QFileInfo(files[0]).baseName().replace("_var_annotated", "");
+		dlg.addFile(name, "VCF", files[0], ui_.actionIgvSample->isChecked());
+	}
 
-		//sample VCF
+	//sample BAM file(s)
+	QList<IgvFile> bams = getBamFiles();
+	foreach(const IgvFile& file, bams)
+	{
+		dlg.addFile(file.id, file.type, file.filename, true);
+	}
 
-		QString folder = QFileInfo(filename_).absolutePath();
-		QStringList files = Helper::findFiles(folder, "*_var_annotated.vcf.gz", false);
+	//sample Manta evidence file(s)
+	QList<IgvFile> evidence_files = getMantaEvidenceFiles();
+	foreach(const IgvFile& file, evidence_files)
+	{
+		dlg.addFile(file.id, file.type, file.filename, false);
+	}
 
-		if (files.count()==1)
+
+	//sample CNV file(s)
+	QList<IgvFile> segs = getSegFilesCnv();
+	foreach(const IgvFile& file, segs)
+	{
+		dlg.addFile(file.id, file.type, file.filename, true);
+	}
+
+	//sample BAF file(s)
+	QList<IgvFile> bafs = getIgvFilesBaf();
+	foreach(const IgvFile& file, bafs)
+	{
+		dlg.addFile(file.id, file.type, file.filename, true);
+	}
+
+	//target region
+	QString roi = ui_.filters->targetRegion();
+	if (roi!="")
+	{
+		dlg.addFile("target region track", "BED", roi, true);
+	}
+
+	//sample low-coverage
+	files = Helper::findFiles(folder, "*_lowcov.bed", false);
+
+	if (files.count()==1)
+	{
+		dlg.addFile("low-coverage regions track", "BED", files[0], ui_.actionIgvLowcov->isChecked());
+	}
+
+	//amplicon file (of processing system)
+	try
+	{
+		NGSD db;
+		ProcessingSystemData system_data = db.getProcessingSystemData(db.processingSystemIdFromProcessedSample(processedSampleName()), true);
+		QString amplicons = system_data.target_file.left(system_data.target_file.length()-4) + "_amplicons.bed";
+		if (QFile::exists(amplicons))
 		{
-			QString name = QFileInfo(files[0]).baseName().replace("_var_annotated", "");
-			dlg.addFile(name, "VCF", files[0], ui_.actionIgvSample->isChecked());
-
+			dlg.addFile("amplicons track (of processing system)", "BED", amplicons, true);
 		}
+	}
+	catch(...) {} //Nothing to do here
 
-		//sample BAM file(s)
-		QList<IgvFile> bams = getBamFiles();
-		if (bams.empty()) return;
-		foreach(const IgvFile& file, bams)
-		{
-			dlg.addFile(file.id, file.type, file.filename, true);	
-		}
+	//custom tracks
+	QList<QAction*> igv_actions = ui_.menuTrackDefaults->findChildren<QAction*>();
+	foreach(QAction* action, igv_actions)
+	{
+		QString text = action->text();
+		if (!text.startsWith("custom track:")) continue;
+		dlg.addFile(text, "custom track", action->toolTip().replace("custom track:", "").trimmed(), action->isChecked());
+	}
 
-		//sample Manta evidence file(s)
-		QList<IgvFile> evidence_files = getMantaEvidenceFiles();
-		foreach(const IgvFile& file, evidence_files)
-		{
-			dlg.addFile(file.id, file.type, file.filename, false);
-		}
+	//execute dialog
+	if (!dlg.exec()) return false;
 
-
-		//sample CNV file(s)
-		QList<IgvFile> segs = getSegFilesCnv();
-		foreach(const IgvFile& file, segs)
-		{
-			dlg.addFile(file.id, file.type, file.filename, true);
-		}
-
-		//sample BAF file(s)
-		QList<IgvFile> bafs = getIgvFilesBaf();
-		foreach(const IgvFile& file, bafs)
-		{
-			dlg.addFile(file.id, file.type, file.filename, true);
-		}
-
-		//target region
-		QString roi = ui_.filters->targetRegion();
-		if (roi!="")
-		{
-			dlg.addFile("target region track", "BED", roi, true);
-		}
-
-		//sample low-coverage
-		files = Helper::findFiles(folder, "*_lowcov.bed", false);
-
-		if (files.count()==1)
-		{
-			dlg.addFile("low-coverage regions track", "BED", files[0], ui_.actionIgvLowcov->isChecked());
-		}
-
-		//amplicon file (of processing system)
-		try
-		{
-			NGSD db;
-			ProcessingSystemData system_data = db.getProcessingSystemData(db.processingSystemIdFromProcessedSample(processedSampleName()), true);
-			QString amplicons = system_data.target_file.left(system_data.target_file.length()-4) + "_amplicons.bed";
-			if (QFile::exists(amplicons))
-			{
-				dlg.addFile("amplicons track (of processing system)", "BED", amplicons, true);
-			}
-		}
-		catch(...) {} //Nothing to do here
-
-		//custom tracks
-		QList<QAction*> igv_actions = ui_.menuTrackDefaults->findChildren<QAction*>();
-		foreach(QAction* action, igv_actions)
-		{
-			QString text = action->text();
-			if (!text.startsWith("custom track:")) continue;
-			dlg.addFile(text, "custom track", action->toolTip().replace("custom track:", "").trimmed(), action->isChecked());
-		}
-
-		//execute dialog
-		if (!dlg.exec()) return;
-
+	QApplication::setOverrideCursor(Qt::BusyCursor);
+	try
+	{
 		if (dlg.initializationAction()==IgvDialog::INIT)
 		{
 			QStringList files_to_load = dlg.filesToLoad();
+			QStringList init_commands;
 			init_commands.append("new");
 			init_commands.append("genome " + Settings::string("igv_genome"));
 
@@ -1310,6 +1313,25 @@ void MainWindow::openInIGV(QString region)
 				}
 			}
 
+			//execute commands
+			bool debug = false;
+			foreach(QString command, init_commands)
+			{
+				if (debug) qDebug() << QDateTime::currentDateTime() << "EXECUTING:" << command;
+				socket.write((command + "\n").toLatin1());
+				bool ok = socket.waitForReadyRead(180000); // 3 min timeout (trios can be slow)
+				QString answer = socket.readAll().trimmed();
+				if (!ok || answer!="OK")
+				{
+					if (debug) qDebug() << QDateTime::currentDateTime() << "FAILED: answer:" << answer << " socket error:" << socket.errorString();
+					THROW(Exception, "Could not execute IGV command '" + command + "'.\nAnswer: " + answer + "\nSocket error:" + socket.errorString());
+				}
+				else
+				{
+					if (debug) qDebug() << QDateTime::currentDateTime() << "DONE";
+				}
+			}
+
 			igv_initialized_ = true;
 		}
 		else if (dlg.initializationAction()==IgvDialog::SKIP_SESSION)
@@ -1320,16 +1342,24 @@ void MainWindow::openInIGV(QString region)
 		{
 			//nothing to do there
 		}
-	}
 
-	//send commands to IGV - init
-	if (!executeIGVCommands(init_commands))
+		QApplication::restoreOverrideCursor();
+
+		return true;
+	}
+	catch(Exception& e)
 	{
-		igv_initialized_ = false;
-	}
+		QApplication::restoreOverrideCursor();
 
-	//send commands to IGV - jump
-	executeIGVCommands(QStringList() << "goto " + region);
+		QMessageBox::warning(this, "Error while initializing IGV", e.message());
+
+		return false;
+	}
+}
+
+void MainWindow::openInIGV(QString region)
+{
+	executeIGVCommands(QStringList() << "goto " + region, true);
 }
 
 void MainWindow::openCustomIgvTrack()
@@ -2710,11 +2740,11 @@ void MainWindow::generateEvaluationSheet()
 	}
 
 	//try to get VariantListInfo from the NGSD
-	EvaluationSheetData evaluation_sheet_data = db.evaluationSheetData(sample_id, false);
-	if (evaluation_sheet_data.ps_id == "")
+	QString ps_id = db.processedSampleId(base_name);
+	EvaluationSheetData evaluation_sheet_data = db.evaluationSheetData(ps_id, false);
+	if (evaluation_sheet_data.ps_id == "") //No db entry found > init
 	{
-		//No db entry found -> set default values
-		evaluation_sheet_data = EvaluationSheetData();
+
 		evaluation_sheet_data.ps_id = db.processedSampleId(base_name);
 		evaluation_sheet_data.dna_rna = db.getSampleData(sample_id).name_external;
 		// make sure reviewer 1 contains name not user id
@@ -2723,7 +2753,6 @@ void MainWindow::generateEvaluationSheet()
 		evaluation_sheet_data.reviewer2 = LoginManager::userName();
 		evaluation_sheet_data.review_date2 = QDate::currentDate();
 	}
-
 
 	//Show VaraintSheetEditDialog
 	EvaluationSheetEditDialog* edit_dialog = new EvaluationSheetEditDialog(this);
@@ -3373,7 +3402,8 @@ void MainWindow::generateReportSomaticRTF()
 		somatic_report_settings_.report_config.setHrdScore(0);
 	}
 
-	SomaticReportDialog dlg(somatic_report_settings_, variants_, cnvs_, somatic_control_tissue_variants_, this); //widget for settings
+	SomaticReportDialog dlg(somatic_report_settings_, cnvs_, somatic_control_tissue_variants_, this); //widget for settings
+
 	if(SomaticRnaReport::checkRequiredSNVAnnotations(variants_))
 	{
 		dlg.enableChoiceReportType(true);
@@ -3542,7 +3572,7 @@ void MainWindow::generateReportGermline()
 		}
 	}
 
-	//check that sample in in NGSD
+	//check that sample is in NGSD
 	NGSD db;
 	QString ps_name = processedSampleName();
 	QString sample_id = db.sampleId(ps_name, false);
@@ -3553,19 +3583,8 @@ void MainWindow::generateReportGermline()
 		return;
 	}
 
-	//check disease information
-	DiseaseInfoWidget* widget = new DiseaseInfoWidget(ps_name, sample_id, this);
-	auto dlg = GUIHelper::createDialog(widget, "Disease information", "", true);
-	if (widget->diseaseInformationMissing() && dlg->exec()==QDialog::Accepted)
-	{
-		db.setSampleDiseaseData(sample_id, widget->diseaseGroup(), widget->diseaseStatus());
-	}
-
-	//set diagnostic status
-	report_settings_.diag_status = db.getDiagnosticStatus(processed_sample_id);
-
 	//show report dialog
-	ReportDialog dialog(report_settings_, variants_, cnvs_, svs_, ui_.filters->targetRegion(),this);
+	ReportDialog dialog(ps_name, report_settings_, variants_, cnvs_, svs_, ui_.filters->targetRegion(),this);
 	if (!dialog.exec()) return;
 
 	//set report type
@@ -3583,9 +3602,6 @@ void MainWindow::generateReportGermline()
 	QList<IgvFile> bams = getBamFiles();
 	if (bams.empty()) return;
 	bam_file = bams.first().filename;
-
-	//update diagnostic status
-	db.setDiagnosticStatus(processed_sample_id, report_settings_.diag_status);
 
 	//show busy dialog
 	busy_dialog_ = new BusyDialog("Report", this);
@@ -3903,7 +3919,16 @@ void MainWindow::importBatch(QString title, QString text, QString table, QString
 int MainWindow::igvPort() const
 {
 	int port = Settings::integer("igv_port");
+
+	//if NGSD is enabled, add the user ID (like that, several users can work on one server)
+	if (LoginManager::active())
+	{
+		port += LoginManager::userId();
+	}
+
+	//if manual override is set, use it
 	if (igv_port_manual>0) port = igv_port_manual;
+
 	return port;
 }
 
@@ -5406,48 +5431,103 @@ void MainWindow::updateVariantDetails()
 	var_last_ = var_current;
 }
 
-bool MainWindow::executeIGVCommands(QStringList commands)
+void MainWindow::executeIGVCommands(QStringList commands, bool init_if_not_done)
 {
-	bool success = true;
-
-	QApplication::setOverrideCursor(Qt::BusyCursor);
+	bool debug = false;
 
 	try
 	{
+		QApplication::setOverrideCursor(Qt::BusyCursor);
+
 		//connect
 		QAbstractSocket socket(QAbstractSocket::UnknownSocketType, this);
-		int igv_port = igvPort();
 		QString igv_host = Settings::string("igv_host");
+		int igv_port = igvPort();
+		if (debug) qDebug() << QDateTime::currentDateTime() << "CONNECTING:" << igv_host << igv_port;
 		socket.connectToHost(igv_host, igv_port);
 		if (!socket.waitForConnected(1000))
 		{
-			THROW(Exception, "Could not connect to IGV at host " + igv_host + " and port " + QString::number(igv_port) + ".\nPlease make sure  IGV is started and the remote control port is enabled:\nView => Preferences => Advanced => Enable port");
+			//show message to user
+			QApplication::restoreOverrideCursor();
+			QMessageBox::information(this, "IGV not running", "IGV is not running on port " + QString::number(igv_port) + ".\nIt will be started now!");
+			QApplication::setOverrideCursor(Qt::BusyCursor);
+
+			if (debug) qDebug() << QDateTime::currentDateTime() << "FAILED - TRYING TO START IGV";
+			igv_initialized_ = false;
+
+			//try to start IGV
+			QString igv_app = Settings::string("igv_app").trimmed();
+			if (igv_app.isEmpty())
+			{
+				THROW(Exception, "Could not start IGV: No settings entry for 'igv_app' found!");
+			}
+			if (!QFile::exists(igv_app))
+			{
+				THROW(Exception, "Could not start IGV: IGV application '" + igv_app + "' does not exist!");
+			}
+			bool started = QProcess::startDetached(igv_app + " --port " + QString::number(igv_port));
+			if (!started)
+			{
+				THROW(Exception, "Could not start IGV: IGV application '" + igv_app + "' did not start!");
+			}
+			if (debug) qDebug() << QDateTime::currentDateTime() << "STARTED IGV - WAITING UNTIL CONNECTING TO THE PORT WORKS";
+
+			//wait for IGV to respond after start
+			bool connected = false;
+			QDateTime max_wait = QDateTime::currentDateTime().addSecs(20);
+			while (QDateTime::currentDateTime() < max_wait)
+			{
+				socket.connectToHost(igv_host, igv_port);
+				if (socket.waitForConnected(1000))
+				{
+					if (debug) qDebug() << QDateTime::currentDateTime() << "CONNECTING TO THE PORT WORKS";
+					connected = true;
+					break;
+				}
+			}
+			if (!connected)
+			{
+				THROW(Exception, "Could not start IGV: IGV application '" + igv_app + "' started, but does not respond!");
+			}
+		}
+		QApplication::restoreOverrideCursor();
+
+		//init if necessary
+		if (!igv_initialized_ && init_if_not_done)
+		{
+			if (debug) qDebug() << QDateTime::currentDateTime() << "INITIALIZING IGV FOR CURRENT SAMPLE!";
+			if (!initializeIvg(socket)) return;
 		}
 
 		//execute commands
+		QApplication::setOverrideCursor(Qt::BusyCursor);
 		foreach(QString command, commands)
 		{
+			if (debug) qDebug() << QDateTime::currentDateTime() << "EXECUTING:" << command;
 			socket.write((command + "\n").toLatin1());
-			socket.waitForReadyRead(180000); // 3 min timeout (trios can be slow)
+			bool ok = socket.waitForReadyRead(180000); // 3 min timeout (trios can be slow)
 			QString answer = socket.readAll().trimmed();
-			if (answer!="OK")
+			if (!ok || answer!="OK")
 			{
-				THROW(Exception, "Could not execute IGV command '" + command + "'.\nAnswer: " + answer);
+				if (debug) qDebug() << QDateTime::currentDateTime() << "FAILED: answer:" << answer << " socket error:" << socket.errorString();
+				THROW(Exception, "Could not execute IGV command '" + command + "'.\nAnswer: " + answer + "\nSocket error:" + socket.errorString());
+			}
+			else
+			{
+				if (debug) qDebug() << QDateTime::currentDateTime() << "DONE";
 			}
 		}
 
 		//disconnect
 		socket.disconnectFromHost();
+		QApplication::restoreOverrideCursor();
 	}
 	catch(Exception& e)
 	{
-		QMessageBox::warning(this, "Error while sending command to IGV:", e.message());
-		success = false;
+		QApplication::restoreOverrideCursor();
+
+		QMessageBox::warning(this, "Error while sending command to IGV", e.message());
 	}
-
-	QApplication::restoreOverrideCursor();
-
-	return success;
 }
 
 void MainWindow::editVariantReportConfiguration(int index)
@@ -5493,6 +5573,7 @@ void MainWindow::editVariantReportConfiguration(int index)
 		ReportVariantDialog dlg(variant.toString(), inheritance_by_gene, var_config, this);
 		dlg.setEnabled(!report_settings_.report_config->isFinalized());
 		if (dlg.exec()!=QDialog::Accepted) return;
+
 
 		//update config, GUI and NGSD
 		report_settings_.report_config->set(var_config);