+++ src/GSvar/ReportDialog.h
@@ -5,6 +5,7 @@
 #include "ReportSettings.h"
 #include "VariantList.h"
 #include "ui_ReportDialog.h"
+#include "NGSD.h"
 
 
 ///Report configutation dialog
@@ -15,7 +16,8 @@ class ReportDialog
 	
 public:
 	///Constructor
-	ReportDialog(ReportSettings& settings, const VariantList& variants, const CnvList& cnvs, const BedpeFile& svs, QString target_region, QWidget* parent = 0);
+	ReportDialog(QString ps, ReportSettings& settings, const VariantList& variants, const CnvList& cnvs, const BedpeFile& svs, QString target_region, QWidget* parent = 0);
+
 	///Returns the report/variant type
 	QString type() const
 	{
@@ -24,19 +26,29 @@ public:
 
 
 protected slots:
+	void checkMetaData();
 	void writeBackSettings();
 	void activateOkButtonIfValid();
 	void initGUI();
-	void updateGUI();
+	void updateVariantTable();
+	void updateCoverageCheckboxStatus();
+
+	void editDiseaseGroupStatus();
+	void editDiseaseDetails();
+	void editDiagnosticStatus();
 
 protected:
 	Ui::ReportDialog ui_;
+	QString ps_;
 	ReportSettings& settings_;
 	const VariantList& variants_;
 	const CnvList& cnvs_;
 	const BedpeFile svs_;
 	QString roi_file_;
 	BedFile roi_;
+	NGSD db_;
+
+	QTableWidgetItem* addTableItem(int row, int col, QString text, bool checkable=false, bool checked_and_not_editable=false);
 };
 
 #endif // REPORTDIALOG_H