+++ src/GSvar/ReportSettings.h
@@ -12,10 +12,9 @@ public:
 	///Default constructor
 	ReportSettings();
 
-	DiagnosticStatusData diag_status; //diagnostic status
-
 	QSharedPointer<ReportConfiguration> report_config; //report configuration
 	QString report_type;
+	QList<QPair<VariantType, int>> selected_variants;
 
 	bool show_coverage_details; //slow low-coverage details
 	int min_depth; //cutoff for low-coverage statistics