+++ src/MappingQC/main.cpp
@@ -30,10 +30,10 @@ public:
 		addFlag("no_cont", "Disables sample contamination calculation, e.g. for tumor or non-human samples.");
 		addFlag("debug", "Enables verbose debug outout.");
 		addEnum("build", "Genome build used to generate the input (needed for contamination only).", true, QStringList() << "hg19" << "hg38", "hg19");
-		addString("ref", "Reference genome for CRAM compression (compulsory for CRAM support).", true);
+		addString("ref","Reference genome for CRAM support (mandatory if CRAM is used).", true);
 
 		//changelog
-		changeLog(2020,  11, 27, "Added Cram support.");
+		changeLog(2020,  11, 27, "Added CRAM support.");
 		changeLog(2018,  7, 11, "Added build switch for hg38 support.");
 		changeLog(2018, 03, 29, "Removed '3exons' flag.");
 		changeLog(2016, 12, 20, "Added support for spliced RNA reads (relevant e.g. for insert size)");