+++ src/SampleSimilarity/main.cpp
@@ -35,10 +35,10 @@ public:
 		addInt("min_cov",  "Minimum coverage to consider a SNP for the analysis (BAM mode).",  true,  30);
 		addInt("max_snps",  "The maximum number of high-coverage SNPs to extract from BAM/CRAM. 0 means unlimited (BAM mode).",  true, 2000);
 		addEnum("build", "Genome build used to generate the input (BAM mode).", true, QStringList() << "hg19" << "hg38", "hg19");
-		addString("ref", "Reference genome for CRAM compression (compulsory for CRAM support).", true);
+		addString("ref", "Reference genome for CRAM support (mandatory if CRAM is used).", true);
 
 		//changelog
-		changeLog(2020,  11, 27, "Added Cram support.");
+		changeLog(2020,  11, 27, "Added CRAM support.");
 		changeLog(2019,  2,  8, "Massive speed-up by caching of variants/genotypes instead of loading them again for each comparison.");
 		changeLog(2018, 11, 26, "Add flag 'skip_multi' to ignore multi-allelic sites.");
 		changeLog(2018,  7, 11, "Added build switch for hg38 support.");