+++ src/cppNGS/VariantType.h
@@ -1,6 +1,9 @@
 #ifndef VARIANTTYPE_H
 #define VARIANTTYPE_H
 
+#include "cppNGS_global.h"
+#include "Exceptions.h"
+
 
 //Variant types
 enum class VariantType
@@ -10,5 +13,9 @@ enum class VariantType
 	SVS, //Structural variants
 	INVALID
 };
+
+//Type to string conversion
+QString CPPNGSSHARED_EXPORT variantTypeToString(VariantType type);
+VariantType CPPNGSSHARED_EXPORT stringToVariantType(QString str);
 
 #endif // VARIANTTYPE_H